Added multi-file upload
Added thumbnail for uploads (pictures)
Added multiple-files upload (all types)
Added drop-to-upload (works with click as well)

+++ deletefile.php
@@ -1,14 +1,16 @@
 <?php
-session_start();
+if (!session_id()) { session_start(); };
 require_once('conf/config.php');
 // if ($isloggedin && $use_login == true) {
-	$deletefile = $_POST['filename'];
-	$fullpath = $userpath.$username.'/'.$deletefile;
+	$deletefile = (isset($_POST['filename']) ? $_POST['filename'] : '');
+	if (!empty($deletefile)) {
+		$fullpath = $userpath.$username.'/'.$deletefile;
 
-	$checkthumbs = explode('/',$deletefile);
-	$thumbs = ($checkthumbs[0] == 'pictures') ? unlink($userpath.$username.'/'.$checkthumbs[0].'/thumbs/'.$checkthumbs[1]) : false;
-	unlink($userpath.$username.'/'.$deletefile);
+		$checkthumbs = explode('/',$deletefile);
+		$thumbs = ($checkthumbs[0] == 'pictures') ? unlink($userpath.$username.'/'.$checkthumbs[0].'/thumbs/'.$checkthumbs[1]) : false;
+		unlink($userpath.$username.'/'.$deletefile);
 
-	echo 'success';
+		echo json_encode(["content"=>"File deleted","infotype"=>"success"]);
+	}
 // }
 ?>
